refactor(docs): extract resolveLink into useDocsResolveLink hook [preview:none]
Address PR #1215 review comment 'should be in proper hooks class' on
doc-viewer.tsx:174.

Moves the endpoint-chain resolution (prop → ChatRuntime.endpoints →
default) + the fetch/JSON-parse pipeline + the try/catch envelope into
a dedicated 'useDocsResolveLink(sourceId, override?)' hook next to the
component's other hooks (use-document-tree, use-scroll-spy).

Benefit: any embedder rendering doc content outside <DocViewer>
(custom markdown renderers, link-resolver previews, etc.) can reuse
the same contract without re-implementing the endpoint chain + error
handling. DocViewer loses 25 lines + 2 imports (useCallback,
useChatRuntime, ResolveLinkResult — all moved into the hook).
